refactor: update import paths for SubmissionWithReplies
- Changed the import path for SubmissionWithReplies in SubmissionItem, SubmissionThread, MyPostsPageClient, and PostsPageClient components to point to the correct types file.
- This enhances code clarity and maintainability by ensuring consistent usage of type definitions across components.
