+++ public/sw.js
@@ -1,5 +1,5 @@
 // Enhanced Service Worker with Aggressive Cache Management and Registration Cleanup
-const CACHE_VERSION = '0.434.3'; // Will be replaced with package.json version during build
+const CACHE_VERSION = '0.434.6'; // Will be replaced with package.json version during build
 const CACHE_NAME = `idling-app-cache-${CACHE_VERSION}`;
 const CACHE_METADATA_NAME = `idling-app-cache-metadata-${CACHE_VERSION}`;
